+++ pytsite/reload/res/assets/js/reload.js
@@ -5,7 +5,7 @@ $(function () {
         var link = $(this);
 
         if (confirm(t('pytsite.reload@confirm_application_reload'))) {
-            pytsite.httpApi.post('reload/reload').done(function() {
+            pytsite.httpApi.post('reload').done(function() {
                 link.parent().text(t('pytsite.reload@app_is_reloading'));
                 setTimeout(function() {
                     location.reload();