+++ pytsite/router/_api.py
@@ -5,27 +5,23 @@ from typing import Dict as _Dict, Union as _Union
 from os import path as _path
 from traceback import format_exc as _format_exc
 from urllib import parse as _urlparse
-from werkzeug.routing import Map as _Map, Rule as _Rule, MapAdapter as _MapAdapter
 from werkzeug.exceptions import HTTPException as _HTTPException
 from werkzeug.contrib.sessions import FilesystemSessionStore as _FilesystemSessionStore
 from pytsite import reg as _reg, logger as _logger, http as _http, util as _util, lang as _lang, tpl as _tpl, \
-    threading as _threading, theme as _theme, setup as _setup, events as _events
+    threading as _threading, theme as _theme, setup as _setup, events as _events, routing as _routing
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 # Routes map
-_routes = _Map()
+_routes = _routing.RulesMap()
 
 # Route path aliases
 _path_aliases = {}
 
 # Session store
 _session_store = _FilesystemSessionStore(path=_reg.get('paths.session'), session_class=_http.session.Session)
-
-# Thread safe map adapters collection
-_map_adapters = {}  # type: _Dict[int, _MapAdapter]
 
 # Thread safe requests collection
 _requests = {}  # type: _Dict[int, _http.request.Request]
@@ -37,25 +33,6 @@ _sessions = {}  # type: _Dict[int, _http.session.Session]
 _no_cache = {}  # type: _Dict[int, bool]
 
 
-class Rule(_Rule):
-    """Routing Rule.
-    """
-
-    def __init__(self, url_path: str, **kwargs):
-        self.call = kwargs.pop('call')
-        self.filters = kwargs.pop('filters', ())
-
-        endpoint = kwargs.get('endpoint')
-        try:
-            _routes.iter_rules(endpoint)
-            raise Exception("Endpoint name '{}' already used.".format(endpoint))
-        except KeyError:
-            super().__init__(url_path, **kwargs)
-
-    def get_rules(self, rules_map):
-        return super().get_rules(rules_map)
-
-
 def get_session_store() -> _FilesystemSessionStore:
     """Get session store.
     """
@@ -63,7 +40,7 @@ def get_session_store() -> _FilesystemSessionStore:
 
 
 def set_request(r: _http.request.Request):
-    """Set request for the current thread.
+    """Set request for current thread.
     """
     _requests[_threading.get_id()] = r
 
@@ -92,54 +69,16 @@ def set_no_cache(status: bool):
     _no_cache[_threading.get_id()] = status
 
 
-def add_rule(path: str, name: str = None, call: str = None, defaults: dict = None, methods=None, filters=None):
+def add_rule(path: str, handler, name: str = None, defaults: dict = None, method: str = 'GET', filters: list = None):
     """Add a rule to the router.
-
-    :param methods: str|tuple|list
     """
-    if not name and not call:
-        raise RuntimeError("Either 'name' or 'call' must be specified.")
+    if not name and isinstance(handler, str):
+        name = handler
 
     if filters is None:
         filters = []
 
-    if isinstance(filters, str):
-        filters = [filters]
-
-    if isinstance(methods, str):
-        methods = (methods,)
-
-    if not isinstance(filters, list) and not isinstance(filters, tuple):
-        raise TypeError('Filters must be a string, list or tuple. {} given.'.format(repr(filters)))
-
-    if not name:
-        name = _util.random_str(32)
-
-    if not call:
-        call = name
-
-    if not defaults:
-        defaults = {}
-
-    defaults['_name'] = name
-    defaults['_call'] = call
-
-    rule = Rule(
-        url_path=path,
-        endpoint=name,
-        call=call,
-        defaults=defaults,
-        methods=methods,
-        filters=filters
-    )
-
-    _routes.add(rule)
-
-    # Bind routes map to corresponding adapter.
-    # It is necessary to get ability to build endpoint URLs before dispatch() call.
-    tid = _threading.get_id()
-    if tid not in _map_adapters:
-        _map_adapters[tid] = _routes.bind(server_name())
+    _routes.add(_routing.Rule(path, handler, name, defaults, method, {'filters': filters}))
 
 
 def add_path_alias(alias: str, target: str):
@@ -178,28 +117,7 @@ def resolve_ep_callable(ep_name: str) -> callable:
 def call_ep(ep_name: str, args: dict = None, inp: dict = None):
     """Call an endpoint.
     """
-    # Try to resolve exact endpoint callable
-    try:
-        ep_name = list(_routes.iter_rules(ep_name))[0].call
-    except KeyError:
-        pass
-
-    if args is not None:
-        if '_call' in args:
-            args['_call_orig'] = args['_call']
-        args['_call'] = ep_name
-
-        if '_name' in args:
-            args['_name_orig'] = args['_name']
-        args['_name'] = ep_name
-
-        return resolve_ep_callable(ep_name)(args, inp)
-
-    elif inp is not None:
-        return resolve_ep_callable(ep_name)(inp)
-
-    else:
-        return resolve_ep_callable(ep_name)()
+    return resolve_ep_callable(ep_name)(args, inp)
 
 
 def dispatch(env: dict, start_response: callable):
@@ -218,14 +136,14 @@ def dispatch(env: dict, start_response: callable):
                                                 status=503, content_type='text/html')
         return wsgi_response(env, start_response)
 
-    _map_adapters[tid] = _routes.bind_to_environ(env)
+    # Creating request context
+    req = _http.request.Request(env)
+    set_request(req)
 
     # Remove trailing slash
-    path_info = _map_adapters[tid].path_info
-    if len(path_info) > 1 and path_info.endswith('/'):
-        redirect_url = _re.sub('/$', '', path_info)
-        if _map_adapters[tid].query_args:
-            redirect_url += '?' + _map_adapters[tid].query_args
+    if req.path != '/' and req.path.endswith('/'):
+        redirect_url = _re.sub('/$', '', req.path)
+        redirect_url += '?' + req.query_string.decode('utf-8') if req.query_string else ''
         return _http.response.Redirect(redirect_url, 301)(env, start_response)
 
     # All requests are cached by default
@@ -234,17 +152,17 @@ def dispatch(env: dict, start_response: callable):
     # Detect language from path
     languages = _lang.langs()
     if len(languages) > 1:
-        if _re.search('^/[a-z]{2}(/|$)', env['PATH_INFO']):
+        if _re.search('^/[a-z]{2}(/|$)', req.path):
             # Extract language code as first two-letters of the path
-            lang_code = env['PATH_INFO'][1:3]
+            lang_code = req.path[1:3]
             try:
                 _lang.set_current(lang_code)
-                env['PATH_INFO'] = env['PATH_INFO'][3:]
-                if not env['PATH_INFO']:
-                    env['PATH_INFO'] = '/'
+                req.path = req.path[3:]
+                if not req.path:
+                    req.path = '/'
                 # If requested language is default, redirect to path without language prefix
                 if lang_code == languages[0]:
-                    return _http.response.Redirect(env['PATH_INFO'], 301)(env, start_response)
+                    return _http.response.Redirect(req.path, 301)(env, start_response)
             except _lang.error.LanguageNotSupported:
                 # If language is not defined, do nothing. 404 will fired in the code below.
                 pass
@@ -253,16 +171,11 @@ def dispatch(env: dict, start_response: callable):
             _lang.set_current(languages[0])
 
     # Notify listeners
-    _events.fire('pytsite.router.pre_dispatch', path_info=env['PATH_INFO'])
+    _events.fire('pytsite.router.pre_dispatch', path_info=req.path)
 
     # Loading path alias, if it exists or use current one
-    env['PATH_INFO'] = _path_aliases.get(env['PATH_INFO'], env['PATH_INFO'])
-
-    # Replace url adapter with modified environment
-    _map_adapters[tid] = _routes.bind_to_environ(env)
-
-    # Creating request context
-    set_request(_http.request.Request(env))
+    if req.path in _path_aliases:
+        req.path = _path_aliases[req.path]
 
     # Shortcuts
     request_input = request().inp
@@ -281,11 +194,14 @@ def dispatch(env: dict, start_response: callable):
         _events.fire('pytsite.router.dispatch')
 
         # Search for rule
-        rule, rule_args = _map_adapters[tid].match(return_rule=True)
+        try:
+            rule = _routes.match(req.path, req.method)
+        except _routing.error.RuleNotFound as e:
+            raise _http.error.NotFound(e)
 
         # Processing rule filters
-        for flt in rule.filters:
-            flt_args = rule_args.copy()
+        for flt in rule.attrs['filters']:
+            flt_args = rule.args
             flt_split = flt.split(':')
             flt_endpoint = flt_split[0]
             if len(flt_split) > 1:
@@ -303,17 +219,17 @@ def dispatch(env: dict, start_response: callable):
 
         # Processing response from handler
         try:
-            response_from_callable = call_ep(rule.call, rule_args, request().inp)
+            handler_resp = resolve_ep_callable(rule.handler)(rule.args, request().inp)
         except ImportError as e:
             raise _http.error.NotFound(e)
 
-        if isinstance(response_from_callable, str):
+        if isinstance(handler_resp, str):
             # Minifying output
             if _reg.get('output.minify'):
-                response_from_callable = _util.minify_html(response_from_callable)
-            wsgi_response.data = response_from_callable
-        elif isinstance(response_from_callable, _http.response.Response):
-            wsgi_response = response_from_callable
+                handler_resp = _util.minify_html(handler_resp)
+            wsgi_response.data = handler_resp
+        elif isinstance(handler_resp, _http.response.Response):
+            wsgi_response = handler_resp
         else:
             wsgi_response.data = ''
 
@@ -402,22 +318,17 @@ def base_path(lang: str = None) -> str:
 def server_name():
     """Get server's name.
     """
-    tid = _threading.get_id()
-    if tid in _map_adapters:
-        name = _map_adapters[tid].server_name
-    else:
-        from pytsite import reg
-        name = reg.get('server_name', 'localhost')
+    r = request()
 
-    return name
+    return r.host if r else _reg.get('server_name', 'localhost')
 
 
 def scheme():
     """Get current URL scheme.
     """
-    tid = _threading.get_id()
+    r = request()
 
-    return _map_adapters[_threading.get_id()].url_scheme if tid in _map_adapters else 'http'
+    return r.scheme if (r and r.scheme) else 'http'
 
 
 def base_url(lang: str = None, query: dict = None):
@@ -485,7 +396,7 @@ def url(s: str, **kwargs) -> str:
         r[5] = fragment
 
     # Adding language suffix (only for relative links as source argument)
-    if not bool(parsed_url[0]) and not strip_lang:
+    if not parsed_url[0] and not strip_lang:
         lang_re = '^/({})/'.format('|'.join(_lang.langs()))
         if not _re.search(lang_re, parsed_url[2]):
             r[2] = str(base_path(lang) + parsed_url[2]).replace('//', '/')
@@ -530,14 +441,10 @@ def current_url(strip_query: bool = False, resolve_alias: bool = True, lang: str
     return r
 
 
-def ep_path(endpoint: str, route_args: dict = None) -> str:
+def ep_path(ep_name: str, args: dict = None) -> str:
     """Get path for an endpoint.
     """
-    tid = _threading.get_id()
-    if tid not in _map_adapters:
-        _map_adapters[tid] = _routes.bind(server_name())
-
-    return _map_adapters[tid].build(endpoint, route_args)
+    return _routes.path(ep_name, args)
 
 
 def ep_url(ep_name: str, route_args: dict = None, **kwargs) -> str: