+++ pytsite/routing/__init__.py
@@ -2,7 +2,7 @@
 """
 from . import _error as error
 from ._rule import Rule
-from ._rule_map import RuleMap
+from ._rules_map import RulesMap
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'